backend/requirements.txt
# Production requirements for SurgiScan Document Processing Microservice
# Updated to use new LandingAI ADE library

# Core FastAPI stack
fastapi
uvicorn[standard]
pydantic
pydantic-settings

# Document processing - NEW LandingAI ADE library
landingai-ade

# AI/ML processing
openai
opencv-python-headless
pillow
numpy

# Database
motor
pymongo

# Authentication and security
python-jose[cryptography]
python-multipart
passlib[bcrypt]
PyJWT

# HTTP clients
requests

# File processing and utilities
python-magic
python-dotenv
python-dateutil
pytz

# Logging and monitoring
python-json-logger
sentry-sdk[fastapi]

# Background processing
celery
redis

# Utilities
psutil
python-slugify

# Production server
gunicorn

# Development tools (optional - remove for production)
pytest
pytest-asyncio
pytest-cov
black
isort
flake8